+++ doc/changelog.md
@@ -2,9 +2,13 @@
 * **0.26.2**
   - Respect DOCKER_CONFIG, KUBECONFIG (#1083)
 
-* **0.26.1**
+* **0.27-SNAPSHOT**
+  - Jump to Java 8 as minimal Java version
+  - Fix NPE in docker:remove-volumes when no volume configuration is given (#1086)
   - Fix NPE when no networks are configured (#1055)
   - Fix Base64 encoding for X-Registry-Auth used for Docker authentication (#1084)
+  - Fix property configuration based's build detection (#1078)
+  - Introduce container name patterns for naming containers (#931)
 
 * **0.26.1** (2018-07-20)
   - Simple Dockerfile triggered also when only a single run section is given